Add secure add/delete endpoint

+++ source/src/main/java/fr/bikemanager/rest/MotorcycleResource.java
@@ -1,10 +1,14 @@
 package fr.bikemanager.rest;
 
+import fr.bikemanager.annotation.Secured;
 import fr.bikemanager.dto.DetailedMotorcycleDto;
 import fr.bikemanager.dto.MotorcycleDto;
+import fr.bikemanager.entity.User;
 import fr.bikemanager.manager.MotorcycleManager;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
+
+import fr.bikemanager.manager.UserManager;
 import fr.bikemanager.rest.MotorcycleResource;
 
 import java.util.List;
@@ -16,18 +20,26 @@ import javax.ws.rs.POST;
 import javax.ws.rs.Path;
 import javax.ws.rs.PathParam;
 import javax.ws.rs.Produces;
+import javax.ws.rs.core.Context;
 import javax.ws.rs.core.MediaType;
+import javax.ws.rs.core.SecurityContext;
 
 @Path("/motorcycles")
 @Produces(MediaType.APPLICATION_JSON)
 @Consumes(MediaType.APPLICATION_JSON)
 public class MotorcycleResource {
 
+    @Context
+    private SecurityContext context;
+
     private MotorcycleManager motorcycleManager;
 
+    private UserManager userManager;
+
     @Autowired
-    public MotorcycleResource(MotorcycleManager motorcycleManager) {
+    public MotorcycleResource(MotorcycleManager motorcycleManager, UserManager userManager) {
         this.motorcycleManager = motorcycleManager;
+        this.userManager = userManager;
     }
 
     @GET
@@ -35,9 +47,13 @@ public class MotorcycleResource {
         return motorcycleManager.getAll();
     }
 
+    @Secured
     @POST
     public void addMotorcycle(DetailedMotorcycleDto motorcycleDto) {
-        motorcycleManager.create(motorcycleDto);
+        // Only admin user are allowed to add motorcycle
+        if (isUserAdmin(context.getUserPrincipal().getName())) {
+            motorcycleManager.create(motorcycleDto);
+        }
     }
 
     @GET
@@ -46,9 +62,24 @@ public class MotorcycleResource {
         return motorcycleManager.getById(id);
     }
 
+    @Secured
     @DELETE
     @Path("/{id}")
     public void deleteMotorcycle(@PathParam("id") int id) {
-        motorcycleManager.deleteById(id);
+        // Only admin are allowed to delete motorcycle
+        if (isUserAdmin(context.getUserPrincipal().getName())) {
+            motorcycleManager.deleteById(id);
+        }
+    }
+
+    /**
+     * Check if the user with the provided auth token is admin
+     *
+     * @param authToken the user auth token
+     * @return true if the user is admin one otherwise false
+     */
+    private boolean isUserAdmin(String authToken) {
+        User user = userManager.getByAuthToken(authToken);
+        return user != null && user.isAdmin();
     }
 }
